refactor: centralize environment GcpMode GcsMode detection
Break circular dependencies between job-store-factory and log-store. Create central zero-dependency env.ts.
