+++ api/app/api/jobs/[id]/aggregate-if-done/route.ts
@@ -28,7 +28,9 @@ import { isWorkerRequest } from '@/lib/auth';
 import { aggregateJobResults } from '@/lib/job-store-factory';
 import { errorResponse } from '@/lib/api-response';
 
-const IS_LOCAL_MODE = !process.env.GOOGLE_CLOUD_PROJECT;
+import { isLocalMode } from '@/lib/env';
+
+const IS_LOCAL_MODE = isLocalMode();
 
 export async function POST(
   req: NextRequest,
