+++ worker_flutter/sentry.properties
@@ -1,0 +1,1 @@
+auth_token=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
